requirements.txt
Django
requests
python-dotenv
yapf
pytest